Let's add our Angular.js dependency, and wire up the two cells. We put an ```ng-app``` attribute on our container div, which tells Angular to compile and parse that element and its children as an Angular app.
We also bind the value of out input cell to the model value ```cell```. We haven't set up a controller with ```ng-controller```, but that's ok &mdash; Angular will use the global ``` rootScope``` bind ```rootScope.cell``` to the our input value. We've also bound the contents of the output element to ```rootScope.cell```, so any changes in the first input element should be mirrored in the output element.