Alright &mdash; let's start evaluating expressions. We're going to do that using Angular's built-in ```$parse``` service.
```$parse``` takes an Angular expression and transforms it into a function. That function takes a scope object, and then evaluates the expression in the context of that scope. In our prior version,
 the ```cell``` expression in ```ng-bind="cell"``` is transformed into something like this: ```function(scope) { return scope['cell'];};```. An expression in something like ```ng-bind="f()"``` would be transformed into something like ```function(scope) { return scope[f]();}```.

Angular uses ```$parse``` to let users create context-sensitive bindings and callbacks in the view, but we're going to use ```$parse``` for our own purposes. We'll parse the contents of the input cell, and turn it into a function we can evaluate.

We do that in our ```compute()``` method with ```$parse($scope.cell)``` to create a function from the expression in the cell.  We call that imediately, passing in our current scope like so: ```$parse($scope.cell)($scope)```. By passing in our ```$scope``` object to the function, we make values on our scope available to the expression, which will be important later.

Due to Angular's change-watching system, any changes to the contents of input cell will cause ```$scope.compute()``` to re-evaluate, and update our div. Try it out with a few arithmetic expressions: e.g., ```52163 / 16604```.