Alright, let's start turning this into a spreadsheet. We'll use ```position: absolute``` to place the input and output on top of each other, and get rid of the Bootstrap CSS. The input is pink, and the output is light purple.
When the cell is not focused, we want to show the computed result of the expression. We do that by placing the output cell on top of the input cell. We also want to show the unevaluated expression when we click on the cell. To do that, we'll pass all click events through the out using ```pointer-event: none```. Any clicks to the output will be passed through to the input cell behind it, allowing the input element to get focus. We'll then use a CSS sibling selector to hide the output when the input is selected like so: ```input:focus .ouput { display: none; }```.