Let's turn this into a spreadsheet!
Instead of having a single cell, we're going to have an object called ```$scope.cells``` that maps names of cells like 'A1' or 'C4' to their current expression.

We also have a list of our columns (```$scope.columns```) and rows (```$scope.rows```). We use ```ng-repeat``` to create a ```<tr>``` element for row, and a ```<td>``` element to each column. Inside that ```<td>``` we have the same input and output div as before, except that they're now bound to a value in the ```$scope.cells``` object.

Because we pass in our ```$scope``` object when evaluating our expressions, you can reference other cells within an expression by calling ```compute``` manually &mdash; try setting the value of cell 'A1' to ```2```, and the value of cell 'A2' to ```compute('A1') * 5```.